+++ components/templates/HeroSectionDesktop/HeroSectionDesktop.tsx
@@ -1,3 +1,4 @@
+"use client"
 import clsx from "clsx"
 import { motion } from "framer-motion"
 import { iSanityWallpaperPaintings } from "lib/models/objects/sanityPainting"
