--- pages/404.tsx
@@ -1,12 +1,0 @@
-import Meta from "components/atoms/Meta/Meta"
-import ErrorPage from "components/pages/ErrorPage"
-import React from "react"
-
-export default function Custom404() {
-  return (
-    <>
-      <Meta />
-      <ErrorPage />
-    </>
-  )
-}
